move publish function to intake module
