+++ rubicon_ml/__init__.py
@@ -1,4 +1,9 @@
-from rubicon_ml.client import (
+from ._version import get_versions
+
+__version__ = get_versions()["version"]
+del get_versions
+
+from rubicon_ml.client import (  # noqa: E402
     Artifact,
     Dataframe,
     Experiment,
@@ -8,11 +13,7 @@ from rubicon_ml.client import (
     Project,
     Rubicon,
 )
-
-from ._version import get_versions
-
-__version__ = get_versions()["version"]
-del get_versions
+from rubicon_ml.intake_rubicon.publish import publish  # noqa: E402
 
 __all__ = [
     "Artifact",
@@ -22,5 +23,6 @@ __all__ = [
     "Metric",
     "Parameter",
     "Project",
+    "publish",
     "Rubicon",
 ]
